fix a third warning about %d and long int
2007-12-19  Havoc Pennington  <hp@redhat.com>

	* src/core/display.c (meta_display_open): fix a third warning
	about %d and long int

	* src/core/delete.c (io_from_ping_dialog): fix another warning
	about long int to %d

	* src/core/compositor.c (meta_compositor_new): fix a warning about
	long int to %d

	* src/core/iconcache.c (meta_read_icons): use
	meta_ui_get_fallback_icons() instead of incorrectly including theme.h

	* src/ui/ui.c (meta_ui_get_fallback_icons): new function


svn path=/trunk/; revision=3492

+++ src/core/compositor.c
@@ -2216,7 +2216,7 @@ meta_compositor_new (MetaDisplay *display)
   compositor = g_new (MetaCompositor, 1);
   compositor->display = display;
 
-  meta_verbose ("Creating %d atoms\n", G_N_ELEMENTS (atom_names));
+  meta_verbose ("Creating %d atoms\n", (int) G_N_ELEMENTS (atom_names));
   XInternAtoms (display->xdisplay, atom_names, G_N_ELEMENTS (atom_names),
                 False, atoms);
   compositor->atom_x_root_pixmap = atoms[0];
